Add files via upload
The purpose is to create a model which on providing the tweets(reviews), to the training, should provide the outcome, whether a particular tweet done by an individual is a positive response or a negative or neutral. All reviews that depict positive response contains the good experience of the traveler with airlines. All reviews that depict negative response contains difficulty faced by the traveler. Neutral responses will be, that is not specific to be considered positive or negative.